heatmap/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0
plotly>=5.0.0
click>=8.0.0